+++ headerkit/__init__.py
@@ -3,7 +3,7 @@
 from headerkit._generate import GenerateResult, generate, generate_all
 from headerkit._ir_json import json_to_header
 from headerkit._populate import PopulateResult, PopulateTarget, populate
-from headerkit.backends import get_backend, is_backend_available, list_backends, reload_backends
+from headerkit.backends import get_backend, is_backend_available, list_backends
 from headerkit.install_libclang import auto_install
 from headerkit.ir import (
     Array,
@@ -66,7 +66,6 @@ __all__ = [
     "get_backend",
     "list_backends",
     "is_backend_available",
-    "reload_backends",
     # Writer Protocol
     "WriterBackend",
     # Writer API